Using again FCCDICTDIR variable
