--- config/common_defaults.py
@@ -1,2 +1,0 @@
-#deffccdicts     = "/afs/cern.ch/work/f/fccsw/public/FCCDicts/"
-deffccdicts     = "/cvmfs/fcc.cern.ch/FCCDicts/"
