+++ examples/FCChh/ttHH/finalSel.py
@@ -1,5 +1,3 @@
-from common_defaults import deffccdicts
-
 #python FCChhAnalyses/FCChh/ttHH/dataframe/finalSel.py 
 import sys, os
 import ROOT
@@ -8,7 +6,7 @@ import ROOT
 baseDir  = "FCChh/ttHH/"
 
 ###Link to the dictonary that contains all the cross section informations etc...
-procDict = os.path.join(os.getenv('FCCDICTSDIR', deffccdicts), '') + "FCC_procDict_fcc_v04.json"
+procDict = 'FCC_procDict_fcc_v04.json'
 
 process_list=['mgp8_pp_tthh_lambda100_5f',
               'mgp8_pp_ttz_5f',
