+++ templates/CMakeLists.txt
