+++ templates/classes.h
